Addressed PR feedback and added README.md
